--- src/test/java/chess/SpringChessConsoleApplicationTests.java
@@ -1,13 +1,0 @@
-package chess;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringChessConsoleApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
